Functions Added in homepage.js
Added the html in homepage.html for the date input and where we will draw the D3 charts. In the java script 4 functions were made displayD3(x), week(), month(), and getData(from, to). These functions are used to get the data from the database which will be sent to another function to generate the D3 charts.